+++ src/pages/people/map.tsx
@@ -5,7 +5,7 @@ import HogMap from 'components/HogMap'
 export default function PeopleMapPage(): JSX.Element {
     return (
         <div data-scheme="primary" className="bg-primary h-full relative">
-            <SEO title="Team map - PostHog" />
+            <SEO title="The hog map" />
             <HogMap />
         </div>
     )
